requirements.txt
# Core
anthropic>=0.40.0        # Claude API SDK
pydantic>=2.0            # Data model validation

# Web Framework
streamlit>=1.30.0        # Dashboard UI

# Visualization
plotly>=5.18.0           # Interactive charts

# Utilities
httpx>=0.25.0            # Async HTTP client
python-dotenv>=1.0.0     # Environment variable management
pyyaml>=6.0              # YAML config parsing
loguru>=0.7.0            # Logging

# Text Processing
jieba>=0.42.1            # Chinese word segmentation (for dedup)
